requirements.txt
asarPy==1.0.1
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
asarPy==1.0.1
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
sqlparse==0.5.3
asarPy==1.0.1
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
